Move read_setup_vars into new setup_vars module
