+++ src/dns/common.rs
@@ -10,8 +10,6 @@
 
 use config::{Config, PiholeFile};
 use regex::Regex;
-use std::io::{self, BufReader};
-use std::io::prelude::*;
 use std::process::{Command, Stdio};
 use util;
 
@@ -24,35 +22,6 @@ pub fn is_valid_domain(domain: &str) -> bool {
     valid_chars_regex.is_match(domain)
         && total_length_regex.is_match(domain)
         && label_length_regex.is_match(domain)
-}
-
-/// Read in a value from setupVars.conf
-pub fn read_setup_vars(entry: &str, config: &Config) -> io::Result<Option<String>> {
-    // Open setupVars.conf
-    let reader = BufReader::new(config.read_file(PiholeFile::SetupVars)?);
-
-    // Check every line for the key (filter out lines which could not be read)
-    for line in reader.lines().filter_map(|item| item.ok()) {
-        // Ignore lines without the entry as a substring
-        if !line.contains(entry) {
-            continue;
-        }
-
-        let mut split = line.split("=");
-
-        // Check if we found the key by checking if the line starts with `entry=`
-        if split.next().map_or(false, |section| section == entry) {
-            return Ok(
-                // Get the right hand side if it exists and is not empty
-                split
-                    .next()
-                    .and_then(|item| if item.len() == 0 { None } else { Some(item) })
-                    .map(|item| item.to_owned())
-            )
-        }
-    }
-
-    Ok(None)
 }
 
 /// Reload Gravity to activate changes in lists
