+++ src/dns/list.rs
@@ -12,7 +12,8 @@ use std::io::prelude::*;
 use std::io::{self, BufReader, BufWriter};
 
 use util;
-use dns::common::{is_valid_domain, read_setup_vars};
+use dns::common::is_valid_domain;
+use setup_vars::read_setup_vars;
 use config::{Config, PiholeFile};
 
 /// Check that the file is a domain list, and return `Error::Unknown` otherwise
